Refactor search runner split for #4106
