+++ changelog.d/unreleased/4106.internal.md
@@ -1,0 +1,17 @@
+---
+category: internal
+issues:
+  - 4106
+affected:
+  - src/CodeIndex/Cli/QueryCommandRunner.cs
+  - src/CodeIndex/Cli/QueryCommandRunner.Search.cs
+  - docs/large-file-decomposition-plan.md
+---
+
+## English
+
+- **Split the search command entry point out of QueryCommandRunner (#4106)** — `RunSearch` now lives in `QueryCommandRunner.Search.cs`, reducing the primary runner hotspot without changing CLI behavior.
+
+## 日本語
+
+- **検索コマンドの入口を QueryCommandRunner から分離しました (#4106)** — `RunSearch` を `QueryCommandRunner.Search.cs` に移し、CLI の挙動を変えずに主要な runner hotspot を小さくしました。
